Atlas: Refactor relvar-storage scan.rs to remove blobs
Extracted the `offset_overflow_tests` and `security_tests` inline modules into `offset_overflow.rs` and `security.rs`. Also separated the MVCC-specific `test_scan_visible_*` tests into their own `scan_visible.rs` file. This resolves the 1200+ line Blob smell in `scan.rs` and improves testing structure.
