+++ .jules/atlas.md
@@ -21,3 +21,6 @@
 ## 2024-05-19 - Sub-modularizing the Database Tests Blob
 **Tangle:** The `relvar-core/src/database/tests.rs` file had grown into a 1,170 line Blob containing 45 inline tests. The God File mixed schema operations, DML, transaction boundaries, and virtual relvar tests.
 **Blueprint:** Removed the `tests.rs` monolith and extracted its contents into a specialized sub-module structure under a new `relvar-core/src/database/tests/` directory. Created explicit sub-modules for `schema.rs`, `data.rs`, `integrity.rs`, `transaction.rs`, and `virtual_relvar.rs`, matching the core database domain responsibilities, along with a `common.rs` for shared helpers.
+## 2026-05-10 - Scan tests extraction Blob
+**Tangle:** The `relvar-storage/src/storage/heap/tests/scan.rs` file had grown into a massive Blob containing over 1,200 lines of tests. It improperly mixed standard tuple scanning tests with MVCC visibility tests and included nested modules (`offset_overflow_tests`, `security_tests`) inline.
+**Blueprint:** Removed the `scan.rs` monolith by extracting inline modules to `offset_overflow.rs` and `security.rs` and pulling MVCC-specific logic into `scan_visible.rs`. This split reduced the `scan.rs` file significantly and separated concerns across testing domains.
