+++ relvar-storage/src/storage/heap/tests/mod.rs
@@ -6,7 +6,10 @@ mod corruption;
 mod delete;
 mod gc;
 mod insert;
+mod offset_overflow;
 mod other;
 mod scan;
+mod scan_visible;
+mod security;
 mod update;
 mod version;
